# Conflicts:
#	source/assets/styles/buildkit.css

+++ initData/progressiveTemplates/liveApps/smartfitLive/src/app/page-body/checkout/checkout.component.ts
@@ -117,7 +117,7 @@ export class CheckoutComponent implements OnInit {
   ePay = false;
   ePayFail = false;
   ePayNull = false;
-  
+  payHereUrl;
   constructor(fb: FormBuilder, private ordersService: OrdersService,
     private shippingService: ShippingService,
     private currencyService: CurrencyService,
@@ -479,14 +479,14 @@ export class CheckoutComponent implements OnInit {
   submit(data, type, note) {
 
     if (type == 'creditcard') {
-      this.makeStripePaymentMethod(data, note);
+      this.makeStripePaymentMethod(data, note, type);
     } else {
-      this.authorizeCreditCardMethod(data, note);
+      this.authorizeCreditCardMethod(data, note, type);
     }
 
   }
 
-  makeStripePaymentMethod(cardInformation, note) {
+  makeStripePaymentMethod(cardInformation, note, type) {
 
     this.showSpinner = true;
     this.hideConfirm = true;
@@ -501,7 +501,7 @@ export class CheckoutComponent implements OnInit {
         this.showSpinner = false;
 
         if (res.status == 'succeeded') {
-          this.orderProcess(note);
+          this.orderProcess(note, type);
         } else {
           this.ePay = false;
           this.ePayNull = false;
@@ -523,7 +523,7 @@ export class CheckoutComponent implements OnInit {
 
   };
 
-  authorizeCreditCardMethod(cardInformation, note) {
+  authorizeCreditCardMethod(cardInformation, note, type) {
 
     this.showSpinner = true;
     this.hideConfirm = true;
@@ -537,7 +537,7 @@ export class CheckoutComponent implements OnInit {
         this.showSpinner = false;
 
         if (res.status == 'ok') {
-          this.orderProcess(note);
+          this.orderProcess(note, type);
         } else if (res.data == "Null Response") {
 
           this.ePay = false;
@@ -572,7 +572,7 @@ export class CheckoutComponent implements OnInit {
       })
   }
 
-  orderProcess(note) {
+  orderProcess(note, type) {
 
     if (note) {
       note = note.trim();
@@ -599,7 +599,8 @@ export class CheckoutComponent implements OnInit {
           'currency': this.dataService.currency,
           'puckupId': null,
           'promotionCode': this.payInfo.promotionCode,
-          'note': note
+          'note': note,
+          'paymentType': type
         };
 
     }
@@ -698,7 +699,8 @@ export class CheckoutComponent implements OnInit {
           'currency': this.dataService.currency,
           'puckupId': null,
           'promotionCode': this.payInfo.promotionCode,
-          'note': note
+          'note': note,
+          'paymentType': 'Cash on delivery'
         };
 
     }
@@ -813,7 +815,8 @@ export class CheckoutComponent implements OnInit {
           'currency': this.dataService.currency,
           'puckupId': null,
           'promotionCode': this.payInfo.promotionCode,
-          'note': note
+          'note': note,
+          'paymentType': 'PayPal'
         };
 
 
@@ -825,7 +828,7 @@ export class CheckoutComponent implements OnInit {
   }
 
   payHere(note) {
-
+    let realHostUrl = encodeURIComponent(window.location.protocol+"//"+window.location.host+"/#/");
     if (note) {
       note = note.trim();
     }
@@ -853,59 +856,63 @@ export class CheckoutComponent implements OnInit {
         'currency': this.dataService.currency,
         'puckupId': null,
         'promotionCode': this.payInfo.promotionCode,
-        'note': note
+        'note': note,
+        'paymentType': 'PayHere',
+        'realHostUrl': realHostUrl,
+        'payHereMerchantId': this.payHereMID
         };
 
     }
 
-    this.http.post(SERVER_URL + "/templatesOrder/savePendingOrder", this.orderDetails,{ responseType: 'json' })
+    this.http.post(SERVER_URL + "/templatesOrder/savePendingOrder", this.orderDetails, { responseType: 'json' })
       .subscribe((orderRes: any) => {
+        this.responce = orderRes.name;
+        this.orderDetails['orderId'] = orderRes.id;
+        this.orderDetails['appId'] = orderRes.appId;
 
-        this.http.post(SERVER_URL + "/templatesInventory/updateInventory", this.payInfo.cart,{ responseType: 'text' })
-          .subscribe((res) => {
-            this.dataService.cart.cartItems = [];
-            this.dataService.cart.cartSize = 0;
-            this.dataService.parentobj.cartSize = this.dataService.cart.cartSize;
-            this.dataService.cart.totalPrice = 0;
-            this.dataService.cart.totalQuantity = 0;
+        if (orderRes.status == 404) {
+          this.showSpinner = false;
+          $('#myModal').modal('show');
+        } else {
+          this.http.post(SERVER_URL + "/templatesInventory/updateInventory", this.payInfo.cart, {responseType: 'text'})
+            .subscribe((res) => {
+                this.dataService.cart.cartItems = [];
+                this.dataService.cart.cartSize = 0;
+                this.dataService.parentobj.cartSize = this.dataService.cart.cartSize;
+                this.dataService.cart.totalPrice = 0;
+                this.dataService.cart.totalQuantity = 0;
 
-            //Pushing into order purchase history
-            let appUser: any = this.localStorageService.get('appLocalStorageUser' + this.appId)
+                //Pushing into order purchase history
+                let appUser: any = this.localStorageService.get('appLocalStorageUser' + this.appId)
 
-            if (appUser) {
-              if (this.localStorageService.get("cart" + appUser.registeredUser)) {
-                this.localStorageService.remove("cart" + appUser.registeredUser);
-              }
-            }else{
-              this.localStorageService.remove("cartUnknownUser");
-            }
+                if (appUser) {
+                  if (this.localStorageService.get("cart" + appUser.registeredUser)) {
+                    this.localStorageService.remove("cart" + appUser.registeredUser);
+                  }
+                } else {
+                  this.localStorageService.remove("cartUnknownUser");
+                }
 
-            let city = this.orderDetails.deliveryCity ? this.orderDetails.deliveryCity : "";
-            let streetNo = this.orderDetails.deliveryNo ? this.orderDetails.deliveryNo : "";
-            let streetName = this.orderDetails.deliveryStreet ? this.orderDetails.deliveryStreet : "";
+                this.http.post(
+                  SERVER_URL + '/mobile/getPayHereForm/', this.orderDetails, {responseType: 'text'})
+                .subscribe((res) => {
+                  this.payHereUrl = res;
+                  $('#payhereProcessModal').modal({backdrop: 'static', keyboard: false});
+                 },
+                 (err) => {
+                    console.log(err);
+                 });
 
-
-            window.location.href=(SERVER_URL + '/mobile/getPayHereForm/?name=' +
-              this.orderDetails.customerName + "&amount=" +
-              this.orderDetails.amount + "&currency=" +
-              this.currency.symbol + "&email=" +
-              this.orderDetails.email + "&telNumber=" +
-              this.orderDetails.telNumber + "&item=" +
-              this.orderDetails.item[0].name + "&address=" +
-              streetNo + " " + streetName  + "&city=" +
-              city + "&appId=" + orderRes.appId +
-              "&orderId=" + orderRes.id + "&payHereMerchantId=" + this.payHereMID);
-          },
-          (err) => {
-            console.log(err);
-          });
-
-
-
+              },
+              (err) => {
+                console.log(err);
+              });
+        }
       },
       (err) => {
         console.log(err);
       });
+  }
 
 
     // var inAppBrowserRef;
@@ -971,6 +978,6 @@ export class CheckoutComponent implements OnInit {
     //   }
 
     // }
-  }
+
 
 }
